requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
groq==0.9.0
instructor==1.3.0
openai==1.30.0
spacy==3.7.4
vaderSentiment==3.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
pytesseract==0.3.10
PyMuPDF==1.24.4
pdfplumber==0.11.0
python-docx==1.1.0
Pillow==10.3.0
celery[redis]==5.3.6
pydantic-settings==2.2.1
python-magic==0.4.27
structlog==24.1.0
python-multipart==0.0.9
httpx==0.27.0
email-validator==2.2.0
